requirements.txt
anvil-uplink
app-settings
paho-mqtt
pip
pyodbc
PyYAML       